Move HttpClient to the common lib (#24)
* Update README.md

* Start splitting out HttpClient WIP

* Move HttpClient into the Util lib

* Repackage

* Fix the build

* Fix the build and tests

* Ok now really fix the tests
