+++ core-android/src/test/java/com/mux/core_android/AndroidDeviceTests.kt
@@ -5,12 +5,12 @@ import android.content.Context
 import android.net.ConnectivityManager
 import android.net.NetworkCapabilities
 import android.os.Build
+import com.mux.android.util.allEqual
 import com.mux.core_android.testdoubles.mockActivity
 import com.mux.core_android.testdoubles.mockConnectivityManager16
 import com.mux.core_android.testdoubles.mockConnectivityManager23
 import com.mux.core_android.testdoubles.mockSharedPrefs
 import com.mux.stats.sdk.muxstats.MuxDataSdk
-import com.mux.stats.sdk.muxstats.util.allEqual
 import org.junit.Assert.*
 import org.junit.Test
 import org.robolectric.annotation.Config
