+++ core-android/src/test/java/com/mux/core_android/testdoubles/FakePlayerBinding.kt
@@ -1,9 +1,9 @@
 package com.mux.core_android.testdoubles
 
+import com.mux.android.util.logTag
 import com.mux.core_android.log
 import com.mux.stats.sdk.muxstats.MuxPlayerAdapter
 import com.mux.stats.sdk.muxstats.MuxStateCollector
-import com.mux.stats.sdk.muxstats.util.logTag
 
 class FakePlayerBinding<Player>(val name: String) : MuxPlayerAdapter.PlayerBinding<Player> {
   override fun bindPlayer(player: Player, collector: MuxStateCollector) {
