+++ mux-kt-utils/build.gradle
@@ -49,7 +49,12 @@ muxDistribution {
 }
 
 dependencies {
+  implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1"
+
   testImplementation 'junit:junit:4.13.2'
+  testImplementation 'androidx.test.ext:junit:1.1.3'
+  testImplementation "io.mockk:mockk:1.12.3"
+  testImplementation 'org.robolectric:robolectric:4.8.1'
   androidTestImplementation 'androidx.test.ext:junit:1.1.4'
   androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
 }
