+++ mux-kt-utils/src/main/java/com/mux/android/util/Util.kt
@@ -1,4 +1,4 @@
-package com.mux.stats.sdk.muxstats.util
+package com.mux.android.util
 
 import kotlin.math.ceil
