+++ mux-kt-utils/src/test/java/com/mux/stats/sdk/muxstats/AbsRobolectricTest.kt
@@ -1,0 +1,12 @@
+package com.mux.stats.sdk.muxstats
+
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+import org.robolectric.annotation.Config
+
+@Config(
+  manifest = Config.NONE,
+)
+@RunWith(RobolectricTestRunner::class)
+abstract class AbsRobolectricTest {
+}
